Enable compatibility with old and new image sharp verions. (Only NET60 and above use the new ImageSharp.) Rebase and Cleanup commits.
