+++ PdfSharpCore/PdfSharpCore.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
+	  <TargetFrameworks>netstandard2.0;netcoreapp3.1;net5.0;net6.0;net7.0;net8.0</TargetFrameworks>
     <CheckEolTargetFramework>false</CheckEolTargetFramework>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
     <Authors>Stefan Steiger and Contributors</Authors>
@@ -17,6 +17,12 @@ PdfSharpCore is a partial port of PdfSharp.Xamarin for .NET Core Additionally Mi
     <summary>PdfSharp for .NET Core</summary>
     <IsTrimmable>true</IsTrimmable>
     <EnableTrimAnalyzer>true</EnableTrimAnalyzer>
+
+	  <AssemblyVersion>1.6.0</AssemblyVersion>
+	  <PackageVersion>1.6.0</PackageVersion>
+	  <FileVersion>1.6.0</FileVersion>
+	  <ProductVersion>1.6.0</ProductVersion>
+	  
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -45,11 +51,23 @@ PdfSharpCore is a partial port of PdfSharp.Xamarin for .NET Core Additionally Mi
   <ItemGroup>
     <Folder Include="Resources\images\" />
   </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="SharpZipLib" Version="1.3.3" />
-    <PackageReference Include="SixLabors.ImageSharp" Version="1.0.4" />
-    <PackageReference Include="SixLabors.Fonts" Version="1.0.0-beta17" />
-  </ItemGroup>
+
+	<Choose>
+		<When Condition="'$(TargetFramework)' == 'net6.0' Or '$(TargetFramework)' == 'net7.0' Or '$(TargetFramework)' == 'net8.0'">
+			<ItemGroup >
+				<PackageReference Include="SharpZipLib" Version="1.4.2" />
+				<PackageReference Include="SixLabors.Fonts" Version="2.0.2" />
+				<PackageReference Include="SixLabors.ImageSharp" Version="3.1.2" />
+			</ItemGroup>
+		</When>
+		<Otherwise>
+			<ItemGroup>
+				<PackageReference Include="SharpZipLib" Version="1.4.2" />
+				<PackageReference Include="SixLabors.Fonts" Version="1.0.0-beta17" />
+				<PackageReference Include="SixLabors.ImageSharp" Version="1.0.4" />
+			</ItemGroup>
+		</Otherwise>
+	</Choose>
 	
   <ItemGroup>
     <None Include="..\LICENSE.md" Pack="true" PackagePath="" />
