+++ PdfSharpCore/Utils/ImageSharpImageSource.cs
@@ -7,40 +7,14 @@ using SixLabors.ImageSharp.Formats;
 using SixLabors.ImageSharp.Formats.Bmp;
 using SixLabors.ImageSharp.Formats.Jpeg;
 using SixLabors.ImageSharp.Formats.Png;
+using SixLabors.ImageSharp.Advanced;
 
 namespace PdfSharpCore.Utils
 {
-    public class ImageSharpImageSource<TPixel> : ImageSource where TPixel : unmanaged, IPixel<TPixel>
+    public partial class ImageSharpImageSource<TPixel> : ImageSource where TPixel : unmanaged, IPixel<TPixel>
     {
 
-        public static IImageSource FromImageSharpImage(Image<TPixel> image, IImageFormat imgFormat, int? quality = 75)
-        {
-            var _path = "*" + Guid.NewGuid().ToString("B");
-            return new ImageSharpImageSourceImpl<TPixel>(_path, image, (int)quality, imgFormat is PngFormat);
-        }
-
-        protected override IImageSource FromBinaryImpl(string name, Func<byte[]> imageSource, int? quality = 75)
-        {
-            var image = Image.Load<TPixel>(imageSource.Invoke(), out IImageFormat imgFormat);
-            return new ImageSharpImageSourceImpl<TPixel>(name, image, (int)quality, imgFormat is PngFormat);
-        }
-
-        protected override IImageSource FromFileImpl(string path, int? quality = 75)
-        {
-            var image = Image.Load<TPixel>(path, out IImageFormat imgFormat);
-            return new ImageSharpImageSourceImpl<TPixel>(path, image, (int) quality, imgFormat is PngFormat);
-        }
-
-        protected override IImageSource FromStreamImpl(string name, Func<Stream> imageStream, int? quality = 75)
-        {
-            using (var stream = imageStream.Invoke())
-            {
-                var image = Image.Load<TPixel>(stream, out IImageFormat imgFormat);
-                return new ImageSharpImageSourceImpl<TPixel>(name, image, (int)quality, imgFormat is PngFormat);
-            }
-        }
-
-        private class ImageSharpImageSourceImpl<TPixel2> : IImageSource where TPixel2 : unmanaged, IPixel<TPixel2>
+        private partial class ImageSharpImageSourceImpl<TPixel2> : IImageSource where TPixel2 : unmanaged, IPixel<TPixel2>
         {
             private Image<TPixel2> Image { get; }
             private readonly int _quality;
@@ -56,11 +30,6 @@ namespace PdfSharpCore.Utils
                 Image = image;
                 _quality = quality;
                 Transparent = isTransparent;
-            }
-
-            public void SaveAsJpeg(MemoryStream ms)
-            {
-                Image.SaveAsJpeg(ms, new JpegEncoder() { Quality = this._quality });
             }
 
             public void Dispose()
